+++ src/components/sidebar/index.ts
@@ -1,0 +1,2 @@
+export * from "./Sidebar.js";
+export { Sidebar } from "./Sidebar.js";